A user can create a channel and become the owner and admin of this channel. He can add other admins and remove them. An admin can mute or ban other users and cancel the mute or ban. If a user is muted, he can see messages but not write messages in this channel. If a user is banned he can't see or write messages. => To do: add a time to mute/ban
A channel can be public (all users can join it), private (users receive an invitation to join the channel) or protected (a password is required to access the channel). A channel can be updated.

A user can send messages directly to another user.
A user can leave a channel. If he is the owner, he can choose to destroy the channel or to give the ownership of the channel to another user.

=> To do: block users